misol

+++ arithmetic_op05.py
@@ -7,5 +7,5 @@
 # Find the division of a and b then assign to "answer".
 a = 72
 b = 9 
-answer = a / b
+answer = int( a / b )
 print(answer)
